project business and dao adjustment

+++ server/controller/src/main/java/ai/starwhale/mlops/domain/project/ProjectMapper.java
@@ -12,7 +12,7 @@ import org.apache.ibatis.annotations.Param;
 
 public interface ProjectMapper {
 
-    Long createProject(ProjectEntity project);
+    int createProject(ProjectEntity project);
 
     int deleteProject(@Param("id") Long id);
